added checks in updated migrations

+++ erp_tech_svcs/db/migrate/upgrade/20120517203052_add_queue_to_delayed_jobs.rb
@@ -1,9 +1,13 @@
 class AddQueueToDelayedJobs < ActiveRecord::Migration
   def self.up
-    add_column :delayed_jobs, :queue, :string
+    unless column_exists?(:delayed_jobs, :queue)
+      add_column :delayed_jobs, :queue, :string
+    end
   end
 
   def self.down
-    remove_column :delayed_jobs, :queue
+    if column_exists?(:delayed_jobs, :queue)
+      remove_column :delayed_jobs, :queue
+    end
   end
 end
